+++ Demo/ScrollingNavbarDemo/ViewControllers/TableViewController.swift
@@ -18,11 +18,12 @@ class TableViewController: ScrollingNavigationViewController, UITableViewDelegat
     super.viewDidLoad()
 
     title = "TableView"
-    tableView.contentInset = UIEdgeInsets(top: 44, left: 0, bottom: 0, right: 0)
+    tableView.contentInset = UIEdgeInsets(top: 44, left: 0, bottom: 100, right: 0)
     toolbar.barTintColor = UIColor(red:0.91, green:0.3, blue:0.24, alpha:1)
     toolbar.tintColor = .white
     navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: nil, action: nil)
     navigationController?.navigationBar.barTintColor = UIColor(red:0.91, green:0.3, blue:0.24, alpha:1)
+    navigationItem.searchController = UISearchController(searchResultsController: nil)
   }
 
   // Enable the navbar scrolling
@@ -45,6 +46,10 @@ class TableViewController: ScrollingNavigationViewController, UITableViewDelegat
     let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
     cell.textLabel?.text = "Row \((indexPath as NSIndexPath).row + 1)"
     return cell
+  }
+  
+  func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+    return "Header"
   }
 
 }